+++ src/Network/Network/Common/NetworkResourceManagerProfile.cs
@@ -2111,6 +2111,11 @@ namespace Microsoft.Azure.Commands.Network
                         dest => dest.ManagementIPConfiguration,
                         opt => opt.MapFrom(src => src.ManagementIpConfiguration)
                     )
+                    // AFC configuration is service-managed and read-only
+                    .ForMember(
+                        dest => dest.AfcConfiguration,
+                        opt => opt.Ignore()
+                    )
                     .AfterMap((src, dest) =>
                     {
                         dest.AdditionalProperties = new Dictionary<string, string>()
@@ -2165,6 +2170,7 @@ namespace Microsoft.Azure.Commands.Network
                 cfg.CreateMap<CNM.PSAzureFirewallPacketCaptureParameters, MNM.FirewallPacketCaptureParameters>();
                 cfg.CreateMap<CNM.PSAzureFirewallAutoscaleConfiguration, MNM.AzureFirewallAutoscaleConfiguration>();
                 cfg.CreateMap<CNM.PSAzureFirewallPacketCaptureResponse, MNM.AzureFirewallPacketCaptureResponse>();
+                cfg.CreateMap<MNM.AfcConfiguration, CNM.PSAzureFirewallAfcConfiguration>();
 
                 // MNM to CNM
                 cfg.CreateMap<MNM.AzureFirewall, CNM.PSAzureFirewall>()